ortools, cplex, and gurobi solvers (#34)
* A few bug fixes and improvements to deep_learning

* Fix indexing bug when forming target from model output

* Added verification to see how the NN performs during test time

* Changed learning rate and max epochs, probably need to link together more nodes to get correct output on lossyPseudoHash_d1

* Changed model to output RV value for each factor's RV and loss function to regress on predicted vs. true value of the RV for each factor

* Allow SymBitVec to use PyTorch tensors but it doens't pass SHA256 test

* Fix Tensor implementation of SymBitVec, all SHA256 tests pass now

* Modify bitwise operations to allow working in continuous domain

* Minor changes

* Implementing linear algebraic solver

* Linear algebra step 3 (factorization) does not work because A+A.T is not invertible

* More experimentation with linear algebra solutions

* Adds SAT-solver using Google's ortools

* Adding some analysis about problem size and solve time

* Added gurobi solver

* Implementing C script to parse datasets

* Add ortools MILP solver

* Remove old sat solver

+++ deep_learning/dataset.py
@@ -21,6 +21,6 @@ class HashReversalDataset(Dataset):
         return len(self.bits)
 
     def __getitem__(self, idx):
-        bits = torch.from_numpy(self.bits[idx])
-        target = torch.from_numpy(self.target[idx])
+        bits = torch.from_numpy(self.bits[idx]).float()
+        target = torch.from_numpy(self.target[idx]).float()
         return bits, target
